Add support for changing the images displayed in the grid
If the number of images changes, everthing is recreated. If the number
of images is unchanged, each image's data is checked for changes, and
only those which have changed are recreated.

ProgressiveImage's update() method allows for alternative
implementations that may have more complex DOM hierarchies and more
data to be selective about what they update.